Update transaction handling to use new insert syntax
In the transaction handler we have to use the `jld-processor/expand` instead of the
`json-ld/expand` because this transaction will get expanded again inside of
`fluree/stage2`. Unfortunately, the output of `json-ld/expand` can't actually be put
back in to another call to `json-ld/expand`.

It is ok to use jld-processor in this context because there are only 5 top-level keys
and the expansion doesn't proceed to nested nodes because top-level keys are @type
@json, or scalar values ("ledger").

We also don't need to do much transaction parsing as it is all handled in
`fluree/stage2`, so we can just pass the whole expanded txn without making modifications
to it.

Also deleted `do-default-context-update` as that function was unused.

Updated all the tests to use the new transaction syntax. Some of the query results don't
have the order guaranteed, so I put the results into a set to compare. In one case (in
policy-test), the transaction was relying on the old upsert semantics, so I had to
introduce a delete clause to maintain the expected state.