Update ledger creation handling to use new insert syntax
This is very similar to the transaction handling changes, with the added twist of
handling defaultContext.

To get this to work expediently I've just put the ledger creation opts under the `opts`
key. This is @type @json and so doesn't get expanded. The map is extracted, the keys
keywordized, and then it is passed without any further processing as the ledger create
opts.

I was able to get rid of the normalize-txn function because nobody uses it anymore.

I also had to change the ref syntax in the basic-query-test. It seems like there's a bug
in json-ld/expand that doesn't know how to deal with @type @id.