remove references to connection type
The connection type should not concern anything outside of the connection
implementation code because the connections should behave exactly the same when
interacting through the connection api/protocol. If they don't behave the same,
then something is wrong with the api/protocol and we should fix that instead of
dispatching on connection type outside of the connection implementation.