add task for running sid migration
This introduces the concept of a task. You can supply a config with a task key whose
value is a map that identifies a task and any arguments it needs. The task is then
dispatched on id with a connection and any arguments supplied. Tasks do not allow the
http server to be running, though once we support consensus again that decision will
need to be revisited. There is an example task config in the resources directory.

The sid migration takes a list of ledger aliases and migrates them one at a time.