consolidate header option handling
Now we support all query opts as headers prefixed with "Fluree-". By handling them in
middleware we can ensure that we never accidentally forget to override supplied
identities with the verified identity. We don't currently support transaction opts but
this would be a useful place to introduce that support in the future.