un-nest nested deftest
All of the error-handling-test assertions were not being run because the deftest was
nested under another deftest. I have moved it to the top level but commented out the
last two assertions as they are not passing and need to be handled by changing the
config parsing in the `db` library. Those will need to be fixed in a follow-up PR as
they are orthogonal to the changes in this one.

I also removed a bunch of redundant/useless comments.