support ledger alias as default query address
We now support /fluree/ledger/<ledger-alias> as a default query address, without
specifying a query operation at the end of the uri.

Since ledger aliases can have slashes in them we can no longer use the heuristic of
using the last slash as the operation delimiter, so we now use `/:`.